Use `steady_clock` for all time related actions.
Using `std::system_clock` for periodic execution opens the code up for
changes in the system time, which is probably not what is desired. A
function for executing a function at a specific point in time would
prefer the use of `std::system_clock` to ensure that the correct local
time is used. For scheduling an event every 300ms it is preferable to
have a steady count.

Futher, the use of `std::high_resolution_clock` is misleading as it
does not refer to the same clock in every implementation. Indeed, the
[standard](wg21.link/std17) (23.17.7.3) says as much:

> Objects of classhigh_resolution_clockrepresent clocks with the
> shortest tick period. `high_resolution_clock` may be a synonym
> for `system_clock` or `steady_clock`.

The following answer on [Stack Overflow](https://bit.ly/36XDXUC) gives
some details about implementation status which can be summarized as:

 * GCC:  `using high_resolution_clock = system_clock;`
 * MSVC: `using high_resolution_clock = steady_clock;`

libc++ is slightly more complex and depends on whether or not
`_LIBCPP_HAS_NO_MONOTONIC_CLOCK` is defined:
 * not defined: `typedef steady_clock high_resolution_clock;`
 * defined:     `typedef system_clock high_resolution_clock;`

This may lead to some problems if this code is run on different
platforms with users expecting consistency.

Signed-off-by: David Brown <d.brown@bigdavedev.com>